complete refactor of recurring code
