+++ src/main/java/seedu/planpal/utility/parser/ParserFactory.java
@@ -3,7 +3,7 @@ package seedu.planpal.utility.parser;
 import seedu.planpal.modes.activities.ActivityManager;
 import seedu.planpal.modes.contacts.ContactManager;
 import seedu.planpal.exceptions.PlanPalExceptions;
-import seedu.planpal.modes.expenses.ExpenseManager;
+import seedu.planpal.modes.expenses.managers.ExpenseManager;
 import seedu.planpal.utility.parser.modeparsers.ActivityParser;
 import seedu.planpal.utility.parser.modeparsers.ContactParser;
 import seedu.planpal.utility.parser.modeparsers.ExpenseParser;
