+++ src/main/java/seedu/planpal/utility/parser/modeparsers/ExpenseParser.java
@@ -3,7 +3,7 @@ package seedu.planpal.utility.parser.modeparsers;
 import seedu.planpal.exceptions.EmptyDescriptionException;
 import seedu.planpal.exceptions.IllegalCommandException;
 import seedu.planpal.exceptions.PlanPalExceptions;
-import seedu.planpal.modes.expenses.ExpenseManager;
+import seedu.planpal.modes.expenses.managers.ExpenseManager;
 import seedu.planpal.modes.expenses.ExpenseModeFunctions;
 import seedu.planpal.utility.Ui;
 import seedu.planpal.utility.filemanager.BackUpManager;
@@ -48,7 +48,9 @@ public class ExpenseParser extends Parser {
             case ADD_COMMAND:
                 description = inputParts[1].trim();
                 if (description.contains(ExpenseModeFunctions.RECURRING_TAG)) {
-                    expenseManager.addRecurringExpense(ExpenseModeFunctions.removeRecurring(description));
+                    expenseManager
+                            .getRecurringManager()
+                            .addRecurringExpense(ExpenseModeFunctions.removeRecurring(description));
                     return true;
                 }
                 expenseManager.addExpense(inputParts[1].trim());
