+++ src/test/java/seedu/planpal/expenses/EditExpenseTest.java
@@ -4,7 +4,7 @@ import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import seedu.planpal.exceptions.EmptyDescriptionException;
 import seedu.planpal.exceptions.PlanPalExceptions;
-import seedu.planpal.modes.expenses.ExpenseManager;
+import seedu.planpal.modes.expenses.managers.ExpenseManager;
 
 import java.io.ByteArrayOutputStream;
 import java.io.PrintStream;
