+++ src/test/java/seedu/planpal/expenses/SetBudgetTest.java
@@ -6,7 +6,7 @@ import seedu.planpal.exceptions.PlanPalExceptions;
 import seedu.planpal.exceptions.expenses.InvalidBudgetException;
 import seedu.planpal.exceptions.expenses.NegativeBudgetException;
 import seedu.planpal.exceptions.expenses.NoBudgetException;
-import seedu.planpal.modes.expenses.ExpenseManager;
+import seedu.planpal.modes.expenses.managers.ExpenseManager;
 
 import java.io.ByteArrayOutputStream;
 import java.io.PrintStream;
