+++ src/test/java/seedu/planpal/expenses/ViewExpenseListTest.java
@@ -3,7 +3,7 @@ package seedu.planpal.expenses;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import seedu.planpal.exceptions.PlanPalExceptions;
-import seedu.planpal.modes.expenses.ExpenseManager;
+import seedu.planpal.modes.expenses.managers.ExpenseManager;
 
 import java.io.ByteArrayOutputStream;
 import java.io.PrintStream;
